+++ src/test/main.ts
@@ -18,6 +18,8 @@ import "./ui.test.js";
 import "./round5.test.js";
 import "./round6.test.js";
 import "./round7.test.js";
+import "./round8.test.js";
+import "./round9.test.js";
 
 import { run } from "./runner.js";
 run();